홈 page 일부 기능들 추상화 작업
