+++ src/utils/events/interval.ts
@@ -1,7 +1,7 @@
 export function interval(t, cb) {
   const intervalArr = [];
 
-  function startInterval(args) {
+  function startInterval(args?) {
     cb(args);
     const interval = setInterval(() => cb(args), t * 1000);
     intervalArr.push(interval);
